Dev01

+++ database/migrations/2018_09_18_191947_create_1537287586_coursesdatas_table.php
@@ -16,7 +16,7 @@ class Create1537287586CoursesdatasTable extends Migration
             Schema::create('coursesdatas', function (Blueprint $table) {
                 $table->increments('id');
                 $table->integer('view')->nullable();
-                $table->integer('progress')->nullable();
+                $table->float('progress')->nullable();
                 $table->integer('rating')->nullable();
                 $table->text('testimonal')->nullable();
