+++ database/migrations/2018_09_18_213943_drop_5ba1466f062e6_coursesdatas_table.php
@@ -26,7 +26,7 @@ class Drop5ba1466f062e6CoursesdatasTable extends Migration
             Schema::create('coursesdatas', function (Blueprint $table) {
                 $table->increments('id');
                 $table->integer('view')->nullable();
-                $table->integer('progress')->nullable();
+                $table->float('progress')->nullable();
                 $table->integer('rating')->nullable();
                 $table->text('testimonal')->nullable();
